odroidhc4: add a marker boot-from-SD-write-to-SPI defconfig for the HC4 based the C4 DT
For reasons I don't understand, using the HC4 DT + SPI + PCI stuff in odroid-hc4_defconfig
causes the HC4 to be unable to talk to the SPI MTD. Since it's very common that an user
booting an Armbian SD card on the HC4 will want to write to the SPI MTD (with intent to
de-infest their machine from Petitboot), Armbian has for a long time shipped the
odroid-c4_defconfig for the HC4 image (usually written to SD card), while shipping the
proper odroid-hc4_defconfig for the SPI version (deployed to MTD by armbian-install).

This patch is thus, not required; we could simply use the c4_defconfig for the SD image.
The Armbian odroidhc4.conf board file has relevant notes on this, but it seems to boil
down to the DT, not any other defconfig changes.