GENIO: configs: genio_700_evk_defconfig: Enlarge decompression buffer
The config CONFIG_SYS_BOOTM_LEN specifies the size of buffer used for
kernel decompression process. The default value (0x4000000, 64MiB) is
not sufficient for decompressing kernel image equal to or larger then
22.5MiB.

Now we increase the value of CONFIG_SYS_BOOTM_LEN to 0x5000000 (80MiB) to
prevent this issue from occurring in the future.

Change-Id: I357cb6ab7dd6c25699a1dbfdc7f2a52dfa339a9c
Signed-off-by: Ramax Lo <ramax.lo@mediatek.com>
Signed-off-by: Macpaul Lin <macpaul.lin@mediatek.com>