GENIO: power: Add mt8391 pwrap and mt6359p pmic support
Add new regulator config: DM_REGULATOR_MT6359.

Modify pwrap driver to add support for MediaTek MT8359 SoC platform.

The MT6359P is a eco version for MT6359 regulator.
For the MT8391 platform, we use the MT6359P as the main PMIC.
The MT6359 and MT6359P have different configurations.
Therefore, on the MT8391 platform, we only provide support for
the MT6359P. If support for the MT6359 PMIC is required, it will
need to be re-implemented.

Change-Id: I033e463bf00ecc4639893565ea3c1d6e750eab08
Signed-off-by: Bo-Chen Chen <rex-bc.chen@mediatek.com>